refactor: simpler obj prop str func handling
