+++ supercycleEngineApp/src/dev/devs.dbd
@@ -14,5 +14,5 @@ function(ioASubBuf)
 device(stringout, CONSTANT, devStrOutCtrl, "ControlStrOut")
 
 ## from devObjPropStringIO.cpp
-device(stringin, INST_IO, devObjPropStringIn, "ObjPropString")
+device(stringin, INST_IO, devStringinObjProp, "StringIOObjProp")
 #device(stringout, INST_IO, devSOFromString, "Obj Prop string")
